Added auth check for add todos from checklist template
