+++ orchestra/todos/decorators.py
@@ -4,8 +4,8 @@ from django.views.decorators.csrf import csrf_exempt
 from jsonview.decorators import json_view
 from rest_framework.decorators import api_view
 # from rest_framework.decorators import authentication_classes
-# from rest_framework.decorators import permission_classes
-# from orchestra.todos import IsAssociatedWithProject
+from rest_framework.decorators import permission_classes
+from orchestra.todos.auth import IsAssociatedWithProject
 
 logger = logging.getLogger(__name__)
 
@@ -24,7 +24,7 @@ def api_endpoint(methods):
     def api_endpoint_decorator(func):
         @csrf_exempt
         @api_view(methods)
-        # @permission_classes((IsAssociatedWithProject,))
+        @permission_classes((IsAssociatedWithProject,))
         @json_view
         @api_exception_logger
         def func_wrapper(*args, **kwargs):
